:coin: make playmarket coins whole integers
Scale the economy x10 (10k starting balance) and store coins as
BigInteger, flooring all coin movements toward the house so balances
and payouts are always whole numbers while LMSR shares stay fractional.

+++ tests/cogs/playmarket/conftest.py
@@ -143,7 +143,7 @@ def reconciles(db):
 
 def set_balance(db, user_id, amount):
     with db.session(PlayerAccount) as session:
-        session.get(PlayerAccount, user_id).balance = Decimal(amount)
+        session.get(PlayerAccount, user_id).balance = int(amount)
         session.commit()
